Se separo la logica de crear
