+++ src/actions/Eventos/index.ts
@@ -1,6 +1,6 @@
+export * from './create-Evento';
 export * from './get-evento-by-slug.action';
 export * from './get-eventos';
-// export * from './getEventos'; // Comentado temporalmente por conflicto de nombres
 export * from './getEventosPorUsuario';
 export * from './getFiltros';
 export * from './setCursos';
